Add family gallery form component with preview cleanup
